fixed: Typo in backup-custom-script.sh

+++ backup-custom-script.sh
@@ -5,11 +5,10 @@ IFS=' '
 for PART_NODEV in $BACKUP_PARTITIONS; do
   local TYPE=`blkid -s TYPE -o value "/dev/${PART_NODEV}"`
   if [ "$TYPE" = "ntfs" ]; then
-    mkdir -p /mnt/windows &&
-    ntfs-3g "$PART_NODEV" /mnt/windows
- 
-    rm -fv /mnt/windows/hiberfile.sys
-    rm -fv /mnt/windows/pagefile.sys
+    if mkdir -p /mnt/windows && ntfs-3g "/dev/${PART_NODEV}" /mnt/windows; then
+      rm -fv /mnt/windows/hiberfile.sys
+      rm -fv /mnt/windows/pagefile.sys
+    fi
 
     umount /mnt/windows
   fi
